Fix TestUnexpectedInitReq failures
If a connection waiting to receive an init gets an unsupported init
message, we buffer an errorFrame and close the connection.
However, the connection is closed before the errorFrame can be sent.
Update Close() in this case to close the sendCh which will close
the network connection after any existing frames in the channel are
sent.