Fix panic on SendSystemError
SendSystemError could cause a panic as it sent errors even if the
channel was Closed. Fix by making sure we only send errors if the
connection is in a valid state, and sending the error to sendCh
while holding a read lock to ensure channel is not closed in that
time.