Recheck connection state after creating exchange.
We use optimistic concurrency when we check the state on a call.
There are 2 states that are tied together:
 - The connection state
 - The number of exchanges
Rather than hold the lock for an extended period of time as we are
creating exchanges, we check the state initially, then create the
exchange. After we create the exchange, we need to re-check the state,
since the state may have changed, we may need to throw away the newly
created exchange.