Fix race condition so TestCloseAfterTimeout passes
The core issue is that we had:
goroutine1 waiting for channel1, which closed channel2 after the wait.
goroutine2 waiting on channel1, or send to channel2

When channel1 is closed, goroutine1 might close channel2, and then
goroutine2 will panic due to the closed channel. In general, if the
timeout closes the sendCh right before goroutine2 does a select, then
the process will crash.

To fix this, we only close sendCh once we are sure there are no
references left. We use wait groups to keep track of goroutines that may
reference sendCh. Timeouts do not decrement this wait group, but still
remove the message exchange from the set, so Close can make progress.