Add test to reproduce hasMoreFragments bug
When frame pooling was enabled, some callReq frames came through
with the "more fragments" flag set, even though it was the last
frame for that call. This turns out to be because deferred bytes
are not always set, but instead, the caller assumes it is already 0.

This test reproduces the bug by using a frame pool that always returns
frames with all bits set. Even if there are no more fragments, the
calls will have the "more fragments" flag set, and so the call times
out since it does not receive any more fragments for that call.