Fix for TestFragmentation failures
TestFragmentation sends a large payload (>5 frames), but since the mex
recvCh is a much smaller buffer (2 frames), if the reader has not
processed the frames before the mex receives a new frame, it would
immediately drop the frame and the test would fail.

Change logic to instead try to send the frame till the timeout.
This means a connection's processing may be stalled if there is is a
slow reader that receives a large number of frames.
If we see this in production (very unlikely), we can increase the mex's
recvCh buffer size.