Clean up ephemeral inbound peers
When a peers' inbound connection is closed, we check whether we can
remove the peer. To remove a peer, it needs to:
 - have 0 inbound and 0 outbound connections
 - be part of no subchannels (scCount == 0)

The peer is removed from the root peer list if the above is true.

Since inbound peers are only added to the root peer list, they have
scCount == 0.