Fix race conditions between ctx.Err() and read err
Add test to cause race condition between a context returning an error
and reading an error frame from the remote side. The test caused a panic
and a Warn log sometimes.

The panic was caused by the frame forwarding removing the message
exchange -- which is incorrect since exchanges are only ever removed by
the reader.

The warn has been downgraded to a Info as it is not a warning to get an
error frame or for a context to timeout.