Fix flaky TestFragmentationSlowReader
The handler context will complete after the caller's context
since the context resolution is more granular than the ttl
sent over tchannel. Since the context may not have expired,
the handler may not get an error when reading arguments.

Wait for the handler context to be Done so we know that there will
be errors on reading arguments.