Retry goroutine verification a couple of times
TestWriteAfterTimeout was flaky as a goroutine was not yet updated
to be runnable, even though it was about to be updated.
We should not rely on the state of all goroutines to be consistent
(e.g. one goroutine may have just completed before one become runnable)

We should retry on leaked frames to avoid flakiness.