Fix panic on SendSystemError on timeout
If the request failed to read, technically, the response should not be
used at all. However, Thrift eeds to be able to send system errors on
thrift parsing errors (which may be caused by non-protocol errors).

Ideally, thrift would tell us whether the transport failed or the
protocol failed, and we'd be able to avoid sending errors unless
the protocol failed, but we need better errors from Thrift for that.