Add basic include support
Add includes to the state, as well as a reference to a map containing
all states.

When goType is called with a type that is package.Type, find the include
that package refers to, and use the state for that include to generate
the type rather than the current state.