Free callReq frames on bad requests
If an inbound server recieved a callReq frame which resulted in an error
(e.g. due to arg1 read error, or unable to find a handler), then the
frame was not beeing freed.

Pass the frame to dispatchInbound and free the frame to avoid leaks.

This issue was detected by the frame pool test.