Avoid race condition between Close and Accept
When Close is called on the listener, the goroutine blocked on Accept
does not immediately error out. It's possible that new incoming
connections will be accepted (at least on Linux). See:
https://github.com/golang/go/issues/13762

We can avoid this by making sure that the server is really closed (e.g. not
accepting new connections) in close_test before trying to make a call.