Fix flaky TestTraceReportingEnabled
Inbound span testing was flaky since the following could happen:
 - call is made successfully
 - handler has finished running, but not yet reported span (e.g. in
   doneSending)
 - test receives the data, and checks the spans

Since the handler hasn't sent the span yet, the test would fail. Fix is
to use a channel to synchronize the reporting with the testing. We wait
up to a second for the span to be reported.

Since we only expect one report (and we're using close which would fail
if we got multiple reports), change calls/spans to single variables.