TestThriftTimeout fixes
If the handler never receives the call, the test times out. Update to
use select and t.Errorf instead.

When running on CI, the test runs a lot slower and may not start the
handler within the timeout. Use the timeout multiplier for the timeout
amounts to avoid this race.