Fix data race on incoming calls/pings and Close
If we receive an incoming call while the connection is active,
but the connection starts closing before we've added the exchange
then we trigger a race between the sendChRef.Add and Wait.

See https://github.com/uber/tchannel-go/issues/262 for more info.

We fix this race by ensuring that anything that may trigger a new
exchange or use sendCh will increment an atomic int, and if the atomic
int is negative, stops processing. We defer a decrement to the atomic
int once the method is complete.

In the Close path, we ensure that we only close the channel after
ensuring that there are no pending methods that may use sendCh.