Introduce testutils.TestServer
Unify a variety of testing conveniences (WithServer, WithVerifiedServer,
and introspection) into a single TestServer type and a WithTestServer
context manager. Migrating to this combined struct will make it easier
to have each test case run with and without an interposed relay.

This commit introduces some temporary code duplication, which will be
resolved once all users of WithVerifiedServer are migrated.