Don't wrap contexts in multiple headerCtx
Propagation of cancel requires a new goroutine only if the parent
context is not recognized by the context.Context library. When we
used headerCtx as a parent, it's not recognized, and a new goroutine
is started. If we instead unwrap the headerCtx (which is only used
to provide helper methods to access headers), then creating a new
context no longer needs to start a new goroutine.