Remove error log if onClosedConnRemoved does not find peer
If some peer P has multiple connections C1 and C2, and both close
at the same time, then we have the following sequence of events.
Each connection is running on a separate goroutine.

C1 calls connectionCloseStateChange, removes itself from the connections
list, and is scheduled out before onClosedConnRemoved

C2 calls connectionCloseStateChange, removes itself from the list
and runs onClosedConnRemoved, which will see that there's no connection
for this peer, and it can be safely removed. So it removes P from the
root peer list

C1 continues to run, goes into onClosedConnRemoved, but doesn't find
P in the root peer list, and logs an error.

Since this can happen and does not actually signify an error, we do not
need to log anything in this scenario.