Fix goroutine leak detector to work consistently on OSX and Linux
Previously, we assumed there was a constant number of background
goroutines and that the first returned goroutine is the current
goroutine. Now we:
 - Get the current goroutine ID more reliably and skip it
 - Skip test functions based on the last function on the stack and the
   goroutine state

Once the known goroutines are filtered, we ensure that there's no other
goroutines. This should fix the flaky frame pool tests and the relay
items left tests since we'll give a chance for goroutines to complete
the work they're doing before verification.