Fix flaky TestConnectTimeout test
Test was flaky, complaining that the relay failed to connect.
The sequence of events was:
 - Start a call with long timeout, but short connect timeout
 - Call fails because of connect timeout
 - Test validates the error, and ends the test, going into verify
 - At this point, relay tries to make a connection to the now shutdown
   server. This fails, since the test closed the server.

The issue was that teh 1ms connect timeout was not enough time for the
relay to connect to the server we'd set up.

Increase the connect timeout so the frame relay has time to connect to
the server, and the timeout occurs while waiting for the init req frame
to be sent.