Add TCP clients that speak TChannel without the TChannel library
Use templates requests/responses for a super-fast TChannel client/server
that can be used to easily saturate a remote server or client.

The templated frames are not hardcoded -- only the location of the ID in
the frame is hardcoded. When the templated object is created, it first
creates a real server/client, puts a recording relay in between, and
makes a call. The recording relay stores all frames in memory.

We then replay these, with the ID substituted to make requests and
responses.

On the request size, we need to do a little more work to capture the
duration taken by the request. We need to record the ID and number of
frames we expect so we know when a response has been fully received.

On the server side, we assume that IDs are strictly increasing and
start sending the response after receiving the first request frame.