Test that all variations of the benchmark client/server work together
Use a matrix test so we know that all combinations of options on a
server will work with all combinations of options on the server.

Each test case will run through with all combinations of server and
client options.

Note: we only test Raw right now, as Thrift is not implemented yet
for the raw TCP server/clients, and raw is what we're currently using to
generate a lot of traffic.