Fix races caused by closing sendCh
We have multiple goroutines possibly writing to sendCh, so closing
sendCh causes races unless we add a ton of synchronization between
the goroutines.

In the relay path, we really don't want to add synchronization overhead
for every frame we're handling. The much simpler option is to avoid
closing sendCh, and instead use a separate channel to signal that the
writeFrames goroutine to end. The sendCh will be cleaned up by the
GC automatically when there's no references left to the channel.

It's possible to simplify (and reduce synchronization) in the standard
call path after this change, which can be done separately if we get a
performance benefit. Since the standard call path does much more than
the relay path, it is likely not going to affect performance as much.