Add call validation, and fix flaky call validation tests
Previously, we only validated that a timeout was set, not that a
servicename was specified when making an outbound call.

Add validation, and do it before we get a connection. This avoids
creating a connection that will not be used.

Previously, we would create a connection, do validation, and close the
connection immediately. On the inbound side, this led to a new
connection that was closed almost immediately, and so sometimes the test
would cause a:
"Failed to add connection to peer" "connection is in an invalid state"

By doing validation before we create a connection we make TestNoTimeout
less flaky, while also returning errors earlier and faster.