Add Peer struct to represent a peer, pass it to stats
We may select a peer as part of some group (e.g. canary), and we want to
track the fact that this is a canary peer through stats. Allow users to
return an assignment as part of the relay.Hosts.Get method and set this
peer with assignment on the stats object.