Add ProtectMemFramePool which ensures frames aren't reused after Release
The protectmem package uses mmap to allocate pages and allows disabling
all access to the page using mprotect.

This allows us to implement a frame pool which disables all access to a
frame once it's been released, which will catch use-after-free style
bugs where we use frames after releasing to the pool.