Fix flaky test TestPeerRemoveClosedConnection (#433)
Test was failing with:
    <autogenerated>:10: Unexpected log: [Warn]: Failed to add connection
    to peer [{service testService} {process testService-52687} {hostPort
    127.0.0.1:52687} {connID 418} {localAddr 127.0.0.1:52687}
    {remoteAddr 127.0.0.1:54541} {remoteHostPort 127.0.0.1:54541}
    {direction inbound} {error tchannel error ErrCodeNetwork: connection
    is in an invalid state}]

The sequence of events was:
p1 creates a connection c1 to p2
once connection is successful, p1 closes c1

On the remote side, p2 has responded to c1 and is about to
mark it active, but the network connection gets closed before
it has been added to the peer's list. When p2 resumes adding
the connection to the peer list, it's now closed due to the underlying
network connection being closed.

We can fix this issue by issuing a ping, which will ensure that the
remote side has processed the connection and has added it to the
peer list before closing the connection.