Fix flaky test TestResponseClosedBeforeRequest (#438)
On the writer side, the last flush of the "closeRequest" byte would fail
with "mex has been shutdown". The following sequence of events was
causing the issue:

 - Flush calls flushFragment, which sends a frame with the "closeRequest" byte
 - The remote side receives the byte, and closes the response
 - Sender receives the last response frame, and shuts down the mex
 - Original goroutine in Flush calls newFragment (since previous
   fragment) has been sent
 - newFragment fails since the mex has been shutdown

Since this is a valid sequence of events, we ignore the error from
Flush when writing out the closeRequest byte. This issue is unlikely
to happen in production, since the network calls to send and receive
frames should take longer than scheduling the goroutine running Flush.