Fix flaky TestPeerScoreOnNewConnection (#440)
Test was failing with "Failed to add connection to peer" due to the
connection being in an invalid state.

Similar to #433, the test ended before the connection was added
on both sides, causing the error log. Fix is similar, to use ping
to ensure the connection is added on both sides rather than ending
the test while the connection is still being added.