Handle host:port mismatches between peer.Connect and remote peer
Currently, if we conect to peer as "a:1" and it sends us an init req
header with host:port set to "b:2", we only add it under the peer "b:2".

After this change, the connection will be added under both peers so that
future calls to "a:1" will be able to use the same connection.

When this happens, we log the host:port mismatch under a debug log.
In future, we may want to emit metrics when this happens.

This information (outboundHostPort vs remoteHostPort) is also exposed
via introspection.