relay: Selected peers should only be added to root peers
Previously, the relay was adding the peer returned from RelayHosts.Get
to the shared peer list, which could lead to unexpected results if the
channel's shared peer list is also used to talk to some other service.

We should never add peers to the shared peer list, if the user wants to
add the peer, they can add it to the shared peer list, but the relay
should only access peers through the root peer list.