relay: Increase max tombstones and TTL
If we receive a response frame after the TTL period, we see an info
log "received frame [...] for outbound message that no longer exists".

We're seeing this often in production, so increase the TTL and allow
storing more tombstones in memory which should reduce the likelihood
of hitting this issue.