Only track active connections in peers
Connections that are closing should only be tracked by the channel.
The peer's connection list is used for scoring the peer and for making
new outbound calls, so we should not be tracking non-active connections
that cannot be used for new outbound calls.

Connections going through shutdown will still be tracked by the channel.