relay: Refactor hosts/stats interface
This refactor does not make any logical changes to the code -- although
it does allow for relay to now forward frames across channels by
returning a peer from a separate channel.

The new interface combines the peer selection and stats interface into a
single RelayCall interface that returns a *Peer rather than a
relay.Peer.

Some advantages of this approach:
 - Avoids a map lookup for converting host:port to *Peer
 - Allows returning peers from separate channels
   This allows our Hyperbahn emulation code to significantly reduce the
   number of channels we will need.
 - No more passing around Zone/Pool in tchannel which does not need to
   know about these concepts.