Reduce flakiness in TestRelayConcurrentCalls
TestRelayConcurrentCalls used the no-library benchmark client which
does not limit concurrency and makes all calls at once. Since we have
numClients = 20, combined with 1000 calls, we essentially make 20,000
calls at once to a single server. The server isn't able to keep up which
could lead to stalling the connection, causing the following log:
"Dropping call due to slow connection to destination."

We should provide a way to control the concurrency limits on the client
side. For now, use the normal client which limits to a single concurrent
call per client.

This will lead to ~20 concurrent calls at any time which should avoid
the test flaking out.