Rewrite thrift-gen test to work without removing vendor
thrift-gen tests did not work with the vendor directory, since the
TChannel code required generated code to satisfy an interface which used
types exposed via vendor, but the generated code was in a temporary
directory, so the types it implemented did not reference the same vendor
directory.

This change creates a fake GOPATH in a temporary directory, symlinks
TChannel and the vendor directory, and creates temporary directories
inside of the this TChannel directory so that TChannel and the generated
code share the same references to vendor.