thrift-gen: Add "namespace go" support. (#559)
Currently, if "a.thrift" includes "b.thrift" but "b.thrift" has a
"namespace go" declaration, then the Apache Thrift generated code
uses the "namespace go" value as the package name. thrift-gen was
not respecting the "namespace go" and so the generated code would
not compile.

This will also help work around the following scenario. If we had:
- b/b.thrift which includes:
    b/shared.thrift
- c/c.thrift which includes
    c/shared.thrift

If a.thrift includes "b/b.thrift", everything is OK. As soon as it tries
to include "c/c.thrift", everything breaks since the "shared" package is
overwritten by the last compiled "shared.thrift". We can work around
this by adding a "namespace go" declaration to one of the shared.thrift
file to use a unique package for each of the shared.thrift files,
without having to rename the file (Which may break a lot of existing
code).

Fixes #101