Only create Connection object for active TChannel connections
Currently, we create a Connection object as soon as we have a TCP
connection, and then we try to add more information (such as remote peer
information) to the logging context after the init handshake. This can
lead to data races since we're modifying the logging context while
other goroutines may be logging messages with the same context.

Instead of using locks for the logging context, instead separate the
handshake phase of connections out of Connection.

Fixes #584