Fix flaky test TestInboundConnection (#588)
TestInboundConnection verifies that the InboundConnection returns the
right connection. However when the test was run with a relay, there was
a race that could cause the test to fail.

We have a server channel and a relay, and the test called the server
through the relay, so we had:
server --C1--> relay --C?--> server

The relay had a race between whether it would use the incoming
connection from the server to send the call back, or creating a new
connection. When the connection was reused, the test passed since the
relay responds with the host:port on inbound connections. However when
the relay creates a new connection, it hides the host:port on outbound
connections, and so the server would see an ephemeral host:port and this
would cause the test to fail:
https://travis-ci.org/uber/tchannel-go/jobs/204481657

Updated the test to use a separate channel for making calls instead of
a single channel calling itself. Also disable relays to avoid races.