Fix flaky TestIntrospectNumConnections
See failure:
https://travis-ci.org/uber/tchannel-go/jobs/205167194

Test pings from new clients to a server, and verifies the number of
connections the server should have. The test is flaky since we don't add
connections to the internal connection list before we start the
read/write goroutines on the connection, so we can end up processing
pings/calls on a connection before we've added the connection to the
list.