Handle parallel incoming connection attempts
One major change as of #591 is to do the handshake synchronously instead
of using a Connection object and the associated message exchanges. This
is fine during the outbound path since the caller is waiting for an
active connection. However on the incoming path, we don't want a single
slow connection to affect other incoming connection attempts.

Instead of performing the incoming connection handshake on the accept
goroutine, start a new goroutine to perform the exchange.