Consistent error logging on init handshake errors
Currently, the inbound side logging may log different messages in
different cases. We may log "Couldn't create new TChannelConnection",
"Failed during connection handshake", or both depending on where the
error occurred in the handshake.

Instead, make this consistent with the outbound side to always only
log "Failed during connection handshake". This also ensures that we send
back an error when we get the wrong message type as the first frame on
an init req.