Make frame parsing more resilient to bad inputs (#621)
Currently, if we recieve a frame that is invalid (E.g., the size is
smaller than the header size which is the minimum), then we end up with
a huge payload size which causes an out-of-bounds panic trying to read
the payload.

Add full unit tests for parsing of the frame, as well as a quick test to
ensure TChannel will not panic on bad frames.