Add tchannel.WithoutHeaders (#637)
This adds a function `tchannel.WithoutHeaders` (name TBD), which hides
TChannel Go-specific information from a Context.

This is needed to alleviate the following issue:

Existing TChannel Go services have been using the standard header
propagation behavior of TChannel Go (incoming headers are copied to
outgoing headers) happily and they start using YARPC+TChannel, or a
library that uses YARPC, to make outgoing calls to a different service.

Their expectation would be to have the headers copied over to that
outgoing call, especially if the underlying transport is TChannel. This
will not be supported in YARPC.

That leaves us with two options: silently ignore TChannel headers, or
reject outgoing requests made with TChannel contexts.

If we silently ignore TChannel headers, service owners who rely on this
behavior will suddenly start seeing hard-to-debug breakages. The number
of support tickets for this will be huge.

A better option is for YARPC to reject TChannel contexts on **all**
transports. YARPC will recommend that users remove the header
information from the context (with the use of `tchannel.WithoutHeaders`)
before passing it to YARPC. In the short-term, this adds a little work
for existing TChannel services but in the long run it saves them a lot
of time in terms of debugging.