Only test go1.7-1.9, lint on latest
Currently, we run tests for go1.5 upwards which takes quite a while on
Travis. go1.7 was released over a year ago, so keeping compatibility
with it as the oldest is reasonable.

As part of this, clean up the versions we run LINT on, and simplify
configuration rather than having the Makefile check the version to run
lint on. Include some minor cleanup to pass lint with latest versions.