Use codecov instead of coveralls
We've moved most of our new Go projects to Codecov which has worked
better than Coveralls. Currently, the Coveralls hook just blocks
merging and doesn't report status. Migrate to codecov and clean up the
Travis configuration so we run lint+coverage in a single run, and don't
do the standard tests in that run.