relay: Create timer after adding item to Items list
We currently create a timer before adding the relay item to the list,
but this has the potential to race. If we create a timer which expires
before we add the item, the timeout will not be able to entomb the item
since we won't find the ID in the items list. If this call never
completes, we'll end up leaking the relay item.