Don't double-log on explicit connection closes (#659)
We are never incrementing closeNetworkCalled, so when a connection is
closed, we log both the original trigger of the close, as well as from
readFrames (usually for "use of closed connection") even though we
explicitly called Close on the network connection. Instead we should
only log info logs from readFrames if it was not caused by an explicit
Close.