Fix flaky TestHealthCheckIntegration
Fixes flaky test seen in:
https://travis-ci.org/uber/tchannel-go/jobs/295494508

The test currently checks if a connection is active after making a ping
request, but there's a chance the connection hasn't yet been closed
(e.g., we are still processing the ping response in the background
goroutine).

We already have logic outside of the for loop to avoid this race, the
test just needed to have less ping respones. When failuresToClose is set
to 1, we only need a single failure for the connection to be closed.