Remove double hostPort field in log
Currently, we add the hostPort field to the channel logger, then add it
again inline in the log call.

This doesn't cause an issue for most users as they use logrus/similar
loggers which use a map for fields. However, when using zap, this ends
up logging the same field multiple times.