introspection: Add inactive connections to output
Tried to debug a stuck connection that was stuck in "startClose", but
unfortunately the introspection output was missing the connection as it
expects connections to be added as part of root peers, but peers only
store references to active connections.

Output inactive connections at the top level of the channel (which is
the only thing that has references to inactive connections).