relay: Create timer after adding item to Items list (#657)
We currently create a timer before adding the relay item to the list,
but this has the potential to race. If we create a timer which expires
before we add the item, the timeout will not be able to entomb the item
since we won't find the ID in the items list. If this call never
completes, we'll end up leaking the relay item.

There's also some other races where a timeout as we fail a relay item
or try to delete it. Fix these races by attempting to stop the timer
first, and if we fail to stop the timer, assume the call will timeout
and let the timeout logic handle cleanup.

Fixes #687, #688