thrift: deal with TStruct serialization failures
*NOTE*: I need feedback on XXX lines.

A thrift server handler can return success with a TStruct that may
return an error when writing a serialized form. This error is ignored,
which cause a invalid third argument to be written (or not written at
all).

We can better handle these kinds of errors earlier by buffering a
serialized TStruct and sending along a system error if the serialization
fails.

Fixes #743.